Models Adjustment

+++ src/main/java/vn/edu/vnuk/fashion/model/Height.java
@@ -3,13 +3,15 @@ package vn.edu.vnuk.fashion.model;
 import javax.validation.constraints.NotNull;
 
 public class Height {
-	private int id;
+	private long id;
+	
 	@NotNull
 	private String lable;
-	public int getId() {
+	
+	public long getId() {
 		return id;
 	}
-	public void setId(int id) {
+	public void setId(long id) {
 		this.id = id;
 	}
 	public String getLable() {
