+++ src/main/java/vn/edu/vnuk/fashion/model/Length.java
@@ -4,14 +4,15 @@ import javax.validation.constraints.NotNull;
 
 public class Length {
 
-	private int id;
+	private long id;
+	
 	@NotNull
 	private String label;
 
-	public int getId() {
+	public long getId() {
 		return id;
 	}
-	public void setId(int id) {
+	public void setId(long id) {
 		this.id = id;
 	}
 	public String getLabel() {
