+++ src/main/java/vn/edu/vnuk/fashion/model/Material.java
@@ -3,14 +3,15 @@ package vn.edu.vnuk.fashion.model;
 import javax.validation.constraints.NotNull;
 
 public class Material {
-	private int id;
+	private long id;
 	
 	@NotNull
 	private String label;
-	public int getId() {
+	
+	public long getId() {
 		return id;
 	}
-	public void setId(int id) {
+	public void setId(long id) {
 		this.id = id;
 	}
 	public String getLabel() {
